requirements.txt
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
keybert>=0.8.0
spacy>=3.7.0
numpy>=1.26.0
scikit-learn>=1.4.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0
nltk>=3.8.0
stopwords>=1.0.0
python-docx>=1.1.0

# spaCy модель для NER
# python -m spacy download en_core_web_sm
